+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 2026-02-01: 1.0.2
 
-* Add reading time and page info to book metadata display
+* Add reading time to book metadata display
+* Improve book info readability
 
 ### 2026-02-01: 1.0.1